# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MSC Backend API Documentation",
        "description": "Interactive Swagger API documentation for MSC Backend",
        "contact": {
            "name": "MSC Support"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Base URL"
        }
    ],
    "paths": {
...
        "/reports/hr-summary": {
            "get": {
                "tags": [
                    "HR Management"
                ],
                "summary": "Get HR reports and analytics summary",
                "operationId": "6614834dc9510af8b4441c9e848b1969",
                "parameters": [
                    {
                        "name": "start_date",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Summary stats"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter ONLY your personal access token (e.g., 1|xxxx...)",
                "scheme": "bearer"
            }
        }
    },
...
}